fix: animation logic
