+++ src/index.ts
@@ -1,6 +1,6 @@
 import ReactNativeModal from './modal';
 
-export {default as ReactNativeModal, OnSwipeCompleteParams} from './modal';
+export {default as ReactNativeModal} from './modal';
 export {
   AnimationEvent,
   Animations,
@@ -11,6 +11,7 @@ export {
   OnOrientationChange,
   GestureResponderEvent,
   ModalProps,
+  OnSwipeCompleteParams,
 } from './modal/types';
 
 export default ReactNativeModal;
