+++ src/modal/index.tsx
@@ -1,35 +1,13 @@
 import * as React from 'react';
-import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
-import {
-  BackHandler,
-  InteractionManager,
-  KeyboardAvoidingView,
-  Modal,
-  Platform,
-  View,
-  useWindowDimensions,
-} from 'react-native';
-import * as animatable from 'react-native-animatable';
+import {Modal, View, useWindowDimensions} from 'react-native';
 import {ModalProps, defaultProps} from './types';
 
-import {buildAnimations, initializeAnimations} from '../utils';
-import Backdrop from './backdrop';
-import {usePanResponder} from './hooks';
+import {initializeAnimations} from '../utils';
+import ReactNativeModalContainer from './container';
 import styles from './modal.style';
-import {Direction} from './types';
 
 // Override default react-native-animatable animations
 initializeAnimations();
-
-export type OnSwipeCompleteParams = {
-  swipingDirection: Direction;
-};
-
-type State = {
-  showContent: boolean;
-  isVisible: boolean;
-  isSwipeable: boolean;
-};
 
 function ReactNativeModal(props: ModalProps) {
   const {height: windowDeviceHeight, width: windowDeviceWidth} =
@@ -74,201 +52,15 @@ function ReactNativeModal(props: ModalProps) {
     ...otherProps
   } = mergedProps;
 
-  const containerProps = otherProps;
-
-  const [isSwipeable, _setIsSwipable] = useState<State['isSwipeable']>(
-    !!swipeDirection,
-  );
-  const [isVisible, setIsVisible] = useState<State['isVisible']>(false);
-  const [showContent, setShowContent] =
-    useState<State['showContent']>(propIsVisible);
-
-  const {animationIn, animationOut} = useMemo(
-    () =>
-      buildAnimations({
-        animationIn: animationOutFromProps!,
-        animationOut: animationInFromProps!,
-      }),
-    [animationInFromProps, animationOutFromProps],
-  );
-
-  const contentRef = useRef<animatable.View>(null);
-  const backdropRef = useRef<animatable.View>(null);
-  const isTransitioning = useRef(false);
-  const interactionHandle = useRef<number | null>(null);
-  const onBackButtonPress = useCallback(() => {
-    if (onBackButtonPressFromProps && isVisible) {
-      onBackButtonPressFromProps();
-      return true;
-    }
-    return false;
-  }, [onBackButtonPressFromProps, isVisible]);
-
-  const {panResponder, pan} = usePanResponder(mergedProps, {
-    backdropRef,
-  });
-
-  const open = useCallback(() => {
-    if (isTransitioning.current) {
-      return;
-    }
-    isTransitioning.current = true;
-
-    backdropRef.current?.transitionTo(
-      {opacity: backdropOpacity},
-      backdropTransitionInTiming,
-    );
-
-    if (isSwipeable) {
-      pan.setValue({x: 0, y: 0});
-    }
-
-    if (contentRef.current) {
-      onModalWillShow && onModalWillShow();
-      if (interactionHandle.current === null) {
-        interactionHandle.current =
-          InteractionManager.createInteractionHandle();
-      }
-      contentRef.current.animate(animationIn, animationInTiming).then(() => {
-        isTransitioning.current = false;
-        if (interactionHandle.current) {
-          InteractionManager.clearInteractionHandle(interactionHandle.current);
-          interactionHandle.current = null;
-        }
-        onModalShow();
-      });
-    }
-  }, [
-    animationIn,
-    animationInTiming,
-    backdropOpacity,
-    backdropTransitionInTiming,
-    isSwipeable,
-    isVisible,
-    onModalShow,
-    onModalWillShow,
-    pan,
-  ]);
-
-  const close = useCallback(() => {
-    if (isTransitioning.current) {
-      return;
-    }
-    isTransitioning.current = true;
-
-    backdropRef.current?.transitionTo(
-      {opacity: 0},
-      backdropTransitionOutTiming,
-    );
-
-    if (contentRef.current) {
-      onModalWillHide();
-      if (interactionHandle.current === null) {
-        interactionHandle.current =
-          InteractionManager.createInteractionHandle();
-      }
-      contentRef.current.animate(animationOut, animationOutTiming).then(() => {
-        isTransitioning.current = false;
-        if (interactionHandle.current) {
-          InteractionManager.clearInteractionHandle(interactionHandle.current);
-          interactionHandle.current = null;
-        }
-        setIsVisible(false);
-        setShowContent(false);
-        onModalHide();
-      });
-    }
-  }, [
-    animationOut,
-    animationOutTiming,
-    backdropTransitionOutTiming,
-    isVisible,
-    onModalHide,
-    onModalWillHide,
-  ]);
-
-  useEffect(
-    function componentDidMount() {
-      BackHandler.addEventListener('hardwareBackPress', onBackButtonPress);
-      return function componentWillUnmount() {
-        BackHandler.removeEventListener('hardwareBackPress', onBackButtonPress);
-        if (interactionHandle.current) {
-          InteractionManager.clearInteractionHandle(interactionHandle.current);
-          interactionHandle.current = null;
-        }
-      };
-    },
-    [open, onBackButtonPress],
-  );
-
-  useEffect(
-    function () {
-      if (backdropRef.current) {
-        backdropRef.current.transitionTo(
-          {opacity: backdropOpacity},
-          backdropTransitionInTiming,
-        );
-      }
-    },
-    [backdropOpacity, backdropTransitionInTiming],
-  );
-
-  useEffect(
-    function getDerivedStateFromProps() {
-      if (isVisible !== propIsVisible) {
-        setIsVisible(propIsVisible);
-        setShowContent(propIsVisible);
-        if (propIsVisible && !isVisible) {
-          open();
-        } else if (!isVisible && isVisible) {
-          close();
-        }
-      }
-    },
-    [propIsVisible, open, close, isVisible],
-  );
-
-  const panHandlers = isSwipeable ? panResponder.panHandlers : {};
-  const panPosition = useNativeDriver
-    ? {
-        transform: pan.getTranslateTransform(),
-      }
-    : pan.getLayout();
-
-  const containerView = (
-    <animatable.View
-      {...panHandlers}
-      ref={contentRef}
-      style={[
-        panPosition,
-        {margin: deviceWidth * 0.05, transform: [{translateY: 0}]},
-        styles.content,
-        style,
-      ]}
-      pointerEvents="box-none"
-      useNativeDriver={useNativeDriver}
-      {...containerProps}>
-      {hideModalContentWhileAnimating && useNativeDriver && !showContent ? (
-        <animatable.View />
-      ) : (
-        children
-      )}
-    </animatable.View>
-  );
-
   // If coverScreen is set to false by the user
   // we render the modal inside the parent view directly
-  if (!coverScreen && isVisible) {
+  if (!coverScreen && propIsVisible) {
     return (
       <View
         pointerEvents="box-none"
-        style={[styles.backdrop, styles.containerBox]}>
-        <Backdrop
-          {...mergedProps}
-          showContent={showContent}
-          ref={backdropRef}
-        />
-        {containerView}
+        style={[styles.backdrop, styles.containerBox]}
+      >
+        <ReactNativeModalContainer {...mergedProps} />
       </View>
     );
   }
@@ -277,24 +69,11 @@ function ReactNativeModal(props: ModalProps) {
     <Modal
       transparent={true}
       animationType={'none'}
-      visible={isVisible}
-      onRequestClose={onBackButtonPress}
-      {...otherProps}>
-      <Backdrop {...mergedProps} showContent={showContent} ref={backdropRef} />
-      {avoidKeyboard ? (
-        <KeyboardAvoidingView
-          behavior={Platform.OS === 'ios' ? 'padding' : undefined}
-          pointerEvents="box-none"
-          style={[
-            styles.content,
-            style,
-            {margin: 0},
-          ]}>
-          {containerView}
-        </KeyboardAvoidingView>
-      ) : (
-        containerView
-      )}
+      visible={propIsVisible}
+      onRequestClose={onBackButtonPressFromProps}
+      {...otherProps}
+    >
+      <ReactNativeModalContainer {...mergedProps} />
     </Modal>
   );
 }
